Add npmignore and prepublish script
The .npmignore will ignore certain files in the tree like configuration
and the gulpfile and allow others like the dist folder which must be
included in the package. Other than that, it should be the same as the
gitignore.

Add a prepublish script to build the library before it is published.