Remove `position` property from Player
Rationale: this property is hard to support for implementing players as
a property they must keep up-to-date because it continuously changes
over time measured in microseconds.

Player#seeked() now takes the absolute position instead of a delta for
consistency with the mpris interface and because we no longer store
position on the player to do this calculation.

The 'seeked' event on the Player now emits with the offset to seek only
for a similar reason.

add Player#getPosition() to be overridden by the user to be called
whenver this property is gotten by the properties interface.